put pivot_control_with_sliders.py into the right folder
